chore(testcases): move types and functions to shared/testcases
